feat: Refactor SequenceEditServerLayer to use editSequence adapter function
